learndjango/requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
black==19.10b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.1.6
django-cors-headers==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
httprunner==3.1.4
hupper==1.10.2
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
ipython==7.20.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.9.5
jsonschema==3.2.0
loguru==0.4.1
Markdown==3.3.3
MarkupSafe==1.1.1
more-itertools==8.7.0
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
parso==0.8.1
PasteDeploy==2.1.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
prompt-toolkit==3.0.14
ptyprocess==0.7.0
py==1.10.0
pydantic==1.8.1
Pygments==2.7.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pyramid==1.10.7
pyrsistent==0.17.3
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.14.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
translationstring==1.4
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
venusian==3.0.0
view==0.1
wcwidth==0.2.5
WebOb==1.8.7
zipp==3.4.0
zope.deprecation==4.4.0
zope.interface==5.2.0
